+++ extensions/readwise-to-tana/src/settings.tsx
@@ -4,15 +4,22 @@ import { Book, Highlight } from './useApi'
 
 export type SettingsValues = {
   author: string
+  authorSupertag: string
   category: string
   coverImageUrl: string
+  highlightColor: string
+  highlightHighlightedAt: string
+  highlightLocation: string
   highlightNote: string
   highlightSupertag: string
-  highlightLocation: string
+  highlightUpdatedAt: string
   id: string
+  readwiseUrl: string
   source: string
   supertag: string
   template: string
+  title: string
+  url: string
 }
 
 type SettingsProps = {
@@ -29,15 +36,17 @@ export default function Settings({ handleSave, template }: SettingsProps) {
     id: '1',
     num_highlights: 2,
     source: 'kindle',
+    source_url: null,
+    highlights_url: 'https://example.com/book/highlights',
     title: 'Take Smart Notes',
     highlights: [
       {
-        text: 'This is a highlight',
+        text: 'Highlight',
         location: 10,
         id: 1,
       },
       {
-        text: 'This is a highlight with a note',
+        text: 'Highlight with a note',
         note: 'This is a note',
         id: 2,
       },
@@ -54,7 +63,7 @@ export default function Settings({ handleSave, template }: SettingsProps) {
     >
       <Form.Description
         title="Source"
-        text="Readwise fields -> Tana fields (leave blank to omit from output). Title is always included."
+        text="Readwise fields -> Tana fields (leave blank to omit from output). The name of the node will always be the source's title."
       />
       <Form.TextField
         id="supertag"
@@ -62,11 +71,30 @@ export default function Settings({ handleSave, template }: SettingsProps) {
         info="# can be omitted"
         storeValue
       />
+      <Form.TextField id="title" title="Title" storeValue />
       <Form.TextField id="author" title="Author" storeValue />
+      <Form.TextField
+        id="authorSupertag"
+        title="Author Supertag"
+        storeValue
+        info="This supertag is added to the author, i.e., #person. # can be omitted."
+      />
       <Form.TextField id="id" title="ID" storeValue />
       <Form.TextField id="category" title="Category" storeValue />
       <Form.TextField id="source" title="Source" storeValue />
       <Form.TextField id="coverImageUrl" title="Cover URL" storeValue />
+      <Form.TextField
+        id="url"
+        title="URL"
+        storeValue
+        info="URL to the source"
+      />
+      <Form.TextField
+        id="readwiseUrl"
+        title="Readwise URL"
+        storeValue
+        info="URL to highlights in Readwise"
+      />
       <Form.Separator />
       <Form.Description
         title="Highlight"
@@ -85,6 +113,13 @@ export default function Settings({ handleSave, template }: SettingsProps) {
         info="A child node with '**Note:** {{note}}' will be added if this is omitted"
         storeValue
       />
+      <Form.TextField id="highlightColor" title="Color" storeValue />
+      <Form.TextField
+        id="highlightHighlightedAt"
+        title="Highlighted At"
+        storeValue
+      />
+      <Form.TextField id="highlightUpdatedAt" title="Updated At" storeValue />
       <Form.Separator />
       <Form.TextArea
         info="This is the compiled template that will be used to generate the note."