+++ extensions/readwise-to-tana/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Readwise to Tana Changelog
 
+## [Add new highlight fields and author supertag] - 2023-02-14
+
+- Add more field mappings for highlight tagging: Updated At, Highlighted At, and Color
+- Add ability to set supertag on author
+
+## [Filtering and new fields] - 2023-02-14
+
+- Add filtering by category to library view
+- Add field mappings for Title, URL, and Readwise URL to template
+
 ## [Initial Version] - 2023-02-09
 
 - List books